Move mod testing to its own file, and include mod pair testing.
